\name{ToyMult}
\docType{data}
\alias{ToyMult}
\title{Toy example data for multivariate error-prone covariates}
\description{
This dataset gives an example data for the illustration of usage of \code{\link{augSIMEX}} function. The dataset contains main data and validation data. The main data contains response variable Y and error-prone covariates. Validation data have both correctly observed data and error-prone covariates. See example 2 of \code{\link{augSIMEX}} for the usage.
}
\usage{data(ToyMult)}
\format{A list of two data frames. Main data, a data frame 1000 observations of 6 variables. Validation data, 500 observations of 8 variables.}
\keyword{datasets}