\name{ToyRepeat}
\docType{data}
\alias{ToyRepeat}
\title{Toy example data for univariate error-prone covariates in repeated measurements case}
\description{
This data set gives an example data for the illustration of usage of \code{\link{augSIMEX}} function in adjusting for the case of repeated measurements error. The dataset contains main data and validation data. The main data includes response variable Y and error-prone covariates. Validation data have both correctly observed data and error-prone covariates. See example 3 of \code{\link{augSIMEX}} for the usage.
}
\usage{data(ToyRepeat)}
\format{A list of two data frames. Main data, a data frame with 1000 observations of 5 variables. Validation data, 500 observations of 4 variables.}
\keyword{datasets}

