\name{ToyUni}
\docType{data}
\alias{ToyUni}
\title{Toy example data for univariate error-prone covariates}
\description{
This data set gives an example data for the illustration of usage of \code{\link{augSIMEX}} function. The dataset contains main data, validation data, and true data. True data correspond to the main data, but the all covariates are corrected observed. The main data and true data contains response variable Y. Both main data and validation data contains the error-prone covariates. Validation data and true data have correctly observed data.
}
\usage{data(ToyUni)}
\format{A list of three data frames. Main data, a data frame 1000 observations of 4 variables. Validation data, 500 observations of 5 variables. True data, 1000 observations of 4 variables. }
\keyword{datasets}